requirements.txt
streamlit
pdfplumber
torch
numpy
transformers
scikit-learn
spacy
scispacy
openai
python-dotenv
gdown